Add fractional movement health cost

+++ scripts/benchmark_normal_play.py
@@ -43,7 +43,7 @@ def parse_args():
     parser.add_argument('--tensor-family-capacity', type=positive_int, default=10)
     parser.add_argument('--tensor-cell-capacity', type=positive_int)
     parser.add_argument('--tensor-static-refill-check-every', type=positive_int, default=100)
-    parser.add_argument('--tensor-health-dtype', choices=tuple(HEALTH_DTYPES), default='int32')
+    parser.add_argument('--tensor-health-dtype', choices=tuple(HEALTH_DTYPES), default='float32')
     parser.add_argument('--tensor-matmul-precision', choices=MATMUL_PRECISIONS, default='high')
     parser.add_argument('--tensor-compile-mode', choices=COMPILE_MODES, default='default')
     parser.add_argument('--tensor-compiled-block-steps', type=positive_int, default=10)
